requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
psycopg2-binary==2.9.9   # PostgreSQL driver
gunicorn==21.2.0         # For running Django in production
django-environ
setuptools               # Required for pkg_resources